todo_app/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary==2.9.2
pillow